Add exec; Fix some bugs...

+++ src/aucont_stop/src/main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <sstream>
 #include <string>
+#include <stdexcept>
 #include <aucont_file.h>
 #include <sys/types.h>
 #include <signal.h>
@@ -11,15 +12,14 @@ void print_usage() {
     std::cout << "USAGE: ./aucont_stop PID [SIGNUM]" << std::endl;
 }
 
-void delete_if_absent(pid_t pid) {
-    if (kill(pid, 0) < 0 && errno == ESRCH) {
-        if (aucont::del_container_pid(pid)) {
-            std::cout << "OK" << std::endl;
-        }
-    }
-}
-
 int main(int argc, char* argv[]) {
+    char path_to_exe[1000];
+    realpath(argv[0], path_to_exe);
+    auto exe_path = std::string(path_to_exe);
+    exe_path = exe_path.substr(0, exe_path.find_last_of("/")) + "/";
+    // preparing aucont common resources path
+    aucont::set_aucont_root(exe_path);
+
     if (argc < 2 || argc > 3) {
         print_usage();
         return 0;
@@ -31,11 +31,19 @@ int main(int argc, char* argv[]) {
         signum = atoi(argv[2]);
     }
 
+    auto pids = aucont::get_containers_pids();
+    if (pids.find(pid) != pids.end() && kill(pid, 0) < 0 && errno == ESRCH) {
+        aucont::del_container_pid(pid);
+        return 0;
+    } 
+
+    std::cout << "SIGNAL = " << signum << std::endl;
+
     if (kill(pid, signum) < 0) {
-        delete_if_absent(pid);
         throw std::runtime_error("Can't send signal [ " + std::string(strerror(errno)) + " ] ");
     }
-    delete_if_absent(pid);
+
+    std::cout << pid << std::endl;
     
     return 0;
 }
